Refactor #104, Feat #96 : 좋아요 수를 조회 방법 리팩토링, 공유 게시판 인기글 조회 API 구현

+++ src/main/java/udodog/goGetterServer/service/sharingboard/SharingBoardLikeService.java
@@ -4,8 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import udodog.goGetterServer.model.dto.DefaultRes;
+import udodog.goGetterServer.model.entity.SharingBoard;
 import udodog.goGetterServer.model.entity.SharingBoardLike;
 import udodog.goGetterServer.repository.SharingBoardLikeRepository;
+import udodog.goGetterServer.repository.SharingBoardRepository;
 
 import java.util.Optional;
 
@@ -14,13 +16,24 @@ import java.util.Optional;
 public class SharingBoardLikeService {
 
     private final SharingBoardLikeRepository sharingBoardLikeRepository;
+    private final SharingBoardRepository sharingBoardRepository;
 
     public DefaultRes likeFeature(Long boardId, Long userId) {
-
+        Optional<SharingBoard> optionalSharingBoard = sharingBoardRepository.findById(boardId);
         Optional<SharingBoardLike> sharingBoardLike = sharingBoardLikeRepository.findByUserId(userId);
+
+        if (optionalSharingBoard.isEmpty()){
+            return DefaultRes.response(HttpStatus.OK.value(),"글이 존재하지 않음");
+        }
+
+        SharingBoard sharingBoard = optionalSharingBoard.get();
 
         if(sharingBoardLike.isPresent()){
             sharingBoardLikeRepository.delete(sharingBoardLike.get());
+            sharingBoard.minusLikeCnt();
+
+            sharingBoardRepository.save(sharingBoard);
+
             return DefaultRes.response(HttpStatus.OK.value(),"좋아요 취소");
         }
         else{
@@ -30,7 +43,9 @@ public class SharingBoardLikeService {
                     .build();
 
             sharingBoardLikeRepository.save(like);
+            sharingBoard.plusLikeCnt();
 
+            sharingBoardRepository.save(sharingBoard);
             return DefaultRes.response(HttpStatus.OK.value(),"좋아요 등록");
         }
